Add working data loaders and run_all_loaders script
